requirements.txt
# Python Backend Dependencies
flask==3.0.0
flask-cors==4.0.0
pandas>=2.2.0
openpyxl>=3.1.0
google-genai>=1.0.0
python-dotenv>=1.0.0
plotly>=5.18.0
werkzeug==3.0.1

# For production deployment
gunicorn==21.2.0
